+++ vnext/codegen/react/components/rnwcore/AndroidSwipeRefreshLayout.g.h
@@ -30,7 +30,8 @@ struct AndroidSwipeRefreshLayoutProps : winrt::implements<AndroidSwipeRefreshLay
        progressBackgroundColor = cloneFromProps->progressBackgroundColor;
        size = cloneFromProps->size;
        progressViewOffset = cloneFromProps->progressViewOffset;
-       refreshing = cloneFromProps->refreshing;  
+       refreshing = cloneFromProps->refreshing;
+       onRefresh = cloneFromProps->onRefresh;  
      }
   }
 
@@ -56,6 +57,10 @@ struct AndroidSwipeRefreshLayoutProps : winrt::implements<AndroidSwipeRefreshLay
   REACT_FIELD(refreshing)
   bool refreshing{};
 
+   // These fields can be used to determine if JS has registered for this event
+  REACT_FIELD(onRefresh)
+  bool onRefresh{false};
+
   const winrt::Microsoft::ReactNative::ViewProps ViewProps;
 };
 
@@ -69,8 +74,8 @@ struct AndroidSwipeRefreshLayoutEventEmitter {
 
   using OnRefresh = AndroidSwipeRefreshLayoutSpec_onRefresh;
 
-  void onRefresh(OnRefresh &value) const {
-    m_eventEmitter.DispatchEvent(L"refresh", [&value](const winrt::Microsoft::ReactNative::IJSValueWriter writer) {
+  void onRefresh(OnRefresh &&value) const {
+    m_eventEmitter.DispatchEvent(L"refresh", [value = std::move(value)](const winrt::Microsoft::ReactNative::IJSValueWriter writer) {
       winrt::Microsoft::ReactNative::WriteValue(writer, value);
     });
   }