+++ packages/@react-native-windows/cli/CHANGELOG.md
@@ -1,8 +1,19 @@
 # Change Log - @react-native-windows/cli
 
-<!-- This log was last generated on Wed, 11 Feb 2026 16:10:55 GMT and should not be manually modified. -->
+<!-- This log was last generated on Sat, 07 Mar 2026 04:08:18 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.81.3
+
+Sat, 07 Mar 2026 04:08:18 GMT
+
+### Patches
+
+- Integrate 0.81.6 (30809111+acoates-ms@users.noreply.github.com)
+- Bump TS target (30809111+acoates-ms@users.noreply.github.com)
+- Bump @react-native-windows/codegen to v0.81.3
+- Bump @react-native-windows/telemetry to v0.81.1
 
 ## 0.81.2